9-sem/Info/etl/requirements.txt
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
PyMySQL==1.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
